egl/null: add support for DRM image format modifiers
This change introduces support for image modifiers to platform_null. In
order for it to create an image with modifiers, it relies on libdrm to
iterate all formats with associated modifiers supported by the display
for the primary drm plane in use.

drmModeFormatModifierBlobIterNext() is added to the DRM api in a different
change and it is not upstream at present.

Internal notes:
 [1] IN_FORMATS blobs are available since kernel 4.14:
  - db1689aa61bd1efb5ce9b896e7aa860a85b7f1b6
  - https://patchwork.freedesktop.org/patch/168543

 [2] the dri image->base.version threshold is 14.
  - Unlike for platform_wayland, where no details were found regarding
  why it's using 15
  - dri_interface.h makes createImageWithModifiers available since
  version 14
  - dri/gbm_dri.c as an example checks for minimum version 14.

Change-Id: I0f7b030f6e1943690692674bf18daabfc153208a
Signed-off-by: Luigi Santivetti <luigi.santivetti@imgtec.com>
Signed-off-by: Randolph Sapp <rs@ti.com>